Rename `ops.internal` sub-package to `ops.graph`
